+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# 0.14.0 unreleased
+# 0.14.0 28.05.2018
 
 ## [deps]
  * Update of `dop_common`